feat: first steps in recipecard components
