+++ .github/pull_request_template.md
@@ -4,6 +4,12 @@
 short description or bulleted list of what the pull request contains
 -->
 
+## Issue
+
+<!--
+short description or bulleted list of what the pull request contains
+-->
+
 ## Background
 
 <!--
